pyproject.toml
[build-system]
requires = ["setuptools >= 61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = 'epicore'
dynamic = ["version"]
dependencies = [
    'biopython>=1.80',
    'click>=8.1',
    'matplotlib>=3.4',
    'numpy>=2',
    'pandas>=2',
    'pyyaml>=6.0.2', 
    'openpyxl>=3.1.3'
]
requires-python = '>= 3.12'
description = 'Compute core epitopes from multiple overlapping peptides.'
readme = 'README.md'
license= {text = 'MIT license'}
keywords = ['peptides', 'epitopes', 'core epitopes']
authors = [{name='Jana Hoffmann', email='epicore_jana@family-hoffmann.de'}]
classifiers = ['Programming Language :: Python :: 3', 'Programming Language :: Python :: 3.12']

[project.urls]
Repository='https://github.com/AG-Walz/epicore'

[project.scripts]
epicore = "epicore_utils.epicore_main:main"

[tool.setuptools.packages.find]
include = ["epicore_utils*"]

[tool.setuptools.dynamic]
version = { attr = "epicore_utils.__version__" }